+++ zenml/steps/evaluator/__init__.py
@@ -12,5 +12,5 @@
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
 
-from zenml.core.steps.evaluator.base_evaluator import BaseEvaluatorStep
-from zenml.core.steps.evaluator.tfma_evaluator import TFMAEvaluator
+from zenml.steps.evaluator.base_evaluator import BaseEvaluatorStep
+from zenml.steps.evaluator.tfma_evaluator import TFMAEvaluator
